requirements.txt
meteomatics
numpy
pandas
influxdb-client
requests
python-dotenv